Remove pointless legacy hashtable scan.
In the beginning, items were not removed from the hash table on class teardown.
Memory leaked.  It was bad.

Then a dumb solution was added.  On each class destruction, scan the whole
table looking for the class value stored in it, and remove what's found.

Later the right answer was added.  Use the info about the class being destroyed,
to lookup just the hash table entry that matters and remove that.

But the dumb solution remained.  We kept scanning a hash table, looking
for something that now we never find, because the right solution has already
removed it.

Until today.  Now finally the pointless dumb solution code is removed.

This is the tale of the "namespaceClasses" table.  That's all I'm looking
at today.  I presume very similar tales still lurk seeking their happy
endings.