[4f2e4ff880] ItclDeleteFunction() was failing to remove the entry from the "procMethods" table during class destruction. This left the entry in place, available for retrieval when random chance in memory re-circulation would bring the same key into play again. Retrieving and using pointers to freed structs is bad.
Fix was to stop the filtering out of some entry removals. They had been
blocked because the legacy path to finding the table had been segfaulting
before. The proper fix though, should not have been to leave the table
polluted, but to use the simpler path to that table which already exists.