Workaround of the incompatibility in Tcl_NewObjectInstance() in Tcl 8.6.15. (See Tcl Ticket 154f0982f2)
Without this workaround, Itcl Bug d0126511d9 re-emerges.

I don't think this is a real solution, but it silences the Itcl test suite.

I think the interface between Itcl 4 and Tcl is cursed to be forever
secretly broken outside narrow ranges of use, which fortunately cover
a large space of utility.

A vote of confidence from another Tcler before merge to Itcl trunk
would be appreciated.