Several tests are examining the child namespaces of an internal implementation namespace of TclOO in order to determine whether an Itcl type has instances.
There may have been a time long ago when the Tcl and Itcl implementations
agreed with that indirect probe.  It must be long in the murky past.
These tests have long been failing to test what they purport to test.

Revised the tests to use [itcl::find] to ask Itcl whether instances are
still present, since that's what the tests claim to be testing.